search line with svg clear button
