+++ src/components/controls/CancelButton.vue
@@ -1,0 +1,25 @@
+<template lang="pug">
+q-btn(v-bind="attrs" @click="onClick" flat dense)
+  q-icon
+    svg(xmlns="http://www.w3.org/2000/svg" viewBox="0 0 6 6")
+      path(d="m 0.0078333,6.2668484 c 0,-0.040351 0.68143212,-0.7558654 1.5142936,-1.5900287 L 3.0364206,3.1601579 1.5020219,1.6233601 C 0.65810263,0.77812095 6.24273e-4,0.05356077 0.04095892,0.01322631 0.08129352,-0.02710861 0.80585371,0.63036987 1.6510928,1.4742892 L 3.1878906,3.0086879 4.7246886,1.4742892 C 5.5699277,0.63036987 6.2944878,-0.02710861 6.3348223,0.01322631 6.3751556,0.05356078 5.7176789,0.77812095 4.8737595,1.6233601 L 3.3393606,3.1601579 4.8737595,4.6969556 C 5.7176789,5.5421947 6.3751574,6.2667548 6.3348223,6.3070898 6.2944891,6.347423 5.5699277,5.6899459 4.7246886,4.8460265 L 3.1878906,3.3116283 1.6712292,4.8259218 C 0.25545458,6.2394859 0.0078333,6.4539894 0.0078333,6.2668484 Z")
+</template>
+
+<script>
+import {defineComponent} from "vue";
+
+export default defineComponent({
+  name: 'cancel-button',
+  emits: ['click'],
+  setup(props, ctx) {
+    function onClick() {
+      ctx.emit('click')
+    }
+
+    return {
+      onClick,
+      attrs: ctx.attrs
+    }
+  }
+})
+</script>
